gratifikasi-ai/pyproject.toml
[project]
name = "gratifikasi-ai"
version = "1.0.0"
description = "AI-powered Gratification Compliance Recommendation System"
requires-python = ">=3.11"

dependencies = [
    # Django stack
    "django>=4.2,<5.0",
    "djangorestframework>=3.14",
    "django-environ>=0.11",
    "psycopg[binary]>=3.1",
    "gunicorn>=21.2",
    # FastAPI stack
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # Async / queue
    "celery[redis]>=5.3",
    "redis>=5.0",
    # AI / ML
    "qdrant-client>=1.9",
    "sentence-transformers>=2.7",
    "transformers>=4.40",
    "torch>=2.3",
    # HTTP / resilience
    "httpx>=0.27",
    "tenacity>=8.3",
    # Logging
    "structlog>=24.1",
    # Settings management
    "pydantic-settings>=2.2",
    # Celery beat scheduler
    "django-celery-beat>=2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.1",
    "pytest-django>=4.8",
]
